% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateToPseudoBulk.R
\name{aggregateToPseudoBulk}
\alias{aggregateToPseudoBulk}
\title{Aggregation of single-cell to pseudobulk data}
\usage{
aggregateToPseudoBulk(
  x,
  assay = NULL,
  sample_id = NULL,
  cluster_id = NULL,
  fun = c("sum", "mean", "median", "prop.detected", "num.detected"),
  scale = FALSE,
  verbose = TRUE,
  BPPARAM = SerialParam(progressbar = verbose)
)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{assay}{character string specifying the assay slot to use as 
input data. Defaults to the 1st available (\code{assayNames(x)[1]}).}

\item{sample_id}{character string specifying which variable to use as sample id}

\item{cluster_id}{character string specifying which variable to use as cluster id}

\item{fun}{a character string.
Specifies the function to use as summary statistic.
Passed to \code{summarizeAssayByGroup2}.}

\item{scale}{logical. Should pseudo-bulks be scaled
with the effective library size & multiplied by 1M?}

\item{verbose}{logical. Should information on progress be reported?}

\item{BPPARAM}{a \code{\link[BiocParallel]{BiocParallelParam}}
object specifying how aggregation should be parallelized.}
}
\value{
a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.
\itemize{
\item{If \code{length(by) == 2}, each sheet (\code{assay}) contains 
  pseudobulks for each of \code{by[1]}, e.g., for each cluster when 
  \code{by = "cluster_id"}. Rows correspond to genes, columns to 
  \code{by[2]}, e.g., samples when \code{by = "sample_id"}}.
\item{If \code{length(by) == 1}, the returned SCE will contain only 
  a single \code{assay} with rows = genes and colums = \code{by}.}}
  
  Aggregation parameters (\code{assay, by, fun, scaled}) are stored in 
  \code{metadata()$agg_pars}, and the number of cells that were aggregated 
  are accessible in \code{int_colData()$n_cells}.
}
\description{
Aggregation of single-cell to pseudobulk data.  Adapted from \code{muscat::aggregateData} and has same syntax and results.  But can be much faster in for \code{SingleCellExperiment} backed by H5AD files
}
\details{
Adapted from \code{muscat::aggregateData} and has same syntax and results.  But can be much faster in for \code{SingleCellExperiment} backed by H5AD files because this summarizes counts using \code{\link[DelayedMatrixStats]{DelayedMatrixStats}}
}
\examples{
library(muscat)

# pseudobulk counts by cluster-sample
data(example_sce)

pb <- aggregateToPseudoBulk(example_sce, 
   cluster_id = 'cluster_id', 
   sample_id = 'sample_id',
   verbose=FALSE)

library(SingleCellExperiment)
assayNames(example_sce)  # one sheet per cluster
head(assay(example_sce)) # n_genes x n_samples

# scaled CPM
cpm <- edgeR::cpm(assay(example_sce))
assays(example_sce)$cpm <- cpm
pb <- aggregateToPseudoBulk(example_sce, 
   assay = "cpm", 
   cluster_id = 'cluster_id', 
   sample_id = 'sample_id',
   scale = TRUE, 
   verbose=FALSE)

head(assay(pb))

# aggregate by cluster only
pb <- aggregateToPseudoBulk(example_sce, cluster_id = "cluster_id", verbose=FALSE)
length(assays(pb)) # single assay
head(assay(pb))    # n_genes x n_clusters

}
\references{
Crowell, HL, Soneson, C, Germain, P-L, Calini, D, 
Collin, L, Raposo, C, Malhotra, D & Robinson, MD: 
On the discovery of population-specific state transitions from 
multi-sample multi-condition single-cell RNA sequencing data. 
\emph{bioRxiv} \strong{713412} (2018). 
doi: \url{https://doi.org/10.1101/713412}
}
\author{
Gabriel Hoffman, Helena L Crowell & Mark D Robinson
}
