% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamlet.R, R/dreamletProcessedData.R,
%   R/vpDF.R
\name{assay,dreamletResult,ANY-method}
\alias{assay,dreamletResult,ANY-method}
\alias{assay,dreamletResult,dreamletResult-method}
\alias{assay,dreamletProcessedData,ANY-method}
\alias{assay,dreamletProcessedData,dreamletProcessedData-method}
\alias{assay,vpDF,ANY-method}
\alias{assay,vpDF,vpDF-method}
\title{Get assay}
\usage{
\S4method{assay}{dreamletResult,ANY}(x, i, withDimnames = TRUE, ...)

\S4method{assay}{dreamletProcessedData,ANY}(x, i, withDimnames = TRUE, ...)

\S4method{assay}{vpDF,ANY}(x, i, withDimnames = TRUE, ...)
}
\arguments{
\item{x}{vpDF object}

\item{i}{number indicating index, or string indicating assay}

\item{withDimnames}{not used}

\item{...}{other arguments}
}
\value{
return ith assay
}
\description{
Get assay

Get assay

Get assays by name
}
