% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellCounts.R
\name{cellCounts}
\alias{cellCounts}
\title{Extract cell counts}
\usage{
cellCounts(x)
}
\arguments{
\item{x}{a \code{SingleCellExperiment}}
}
\value{
matrix of cell counts with samples as rows and cell types as columns
}
\description{
Extract matrix of cell counts from \code{SingleCellExperiment}
}
