% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTypeCompositionTest.R
\name{cellTypeCompositionTest}
\alias{cellTypeCompositionTest}
\title{Test difference in cell type composition}
\usage{
cellTypeCompositionTest(
  obj,
  formula,
  coef,
  method = c("nb", "binomial", "betabinomial", "lm", "lmlog", "poisson", "dream")
)
}
\arguments{
\item{obj}{\code{SingleCellExperiment} from \code{aggregateToPseudoBulk}}

\item{formula}{formula indicating variables to include in the regression}

\item{coef}{which coefficient to test}

\item{method}{which regression model to use}
}
\description{
Test difference in cell type composition using a regression models
}
\details{
Evaluates regression model where cell fraction is response using one of the 6 specified methods.  Methods can be fit with random effects, except for 'betabinomial'
}
