% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTypeCompositionTest.R
\name{cellTypeCompositionVarPart}
\alias{cellTypeCompositionVarPart}
\title{Variance partitioning of cell type composition}
\usage{
cellTypeCompositionVarPart(
  obj,
  formula,
  coef,
  method = c("nb", "binomial", "betabinomial", "lm", "lmlog", "poisson")
)
}
\arguments{
\item{obj}{\code{SingleCellExperiment} from \code{aggregateToPseudoBulk}}

\item{formula}{formula indicating variables to include in the regression}

\item{coef}{which coefficient to test}

\item{method}{which regression model to use}
}
\description{
Variance partitioning of cell type composition using a regression models
}
\details{
Evaluates regression model where cell fraction is response using one of the 6 specified methods.  Methods can be fit with random effects, except for 'betabinomial'
}
