% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{checkFormula}
\alias{checkFormula}
\title{Check variables in a formula}
\usage{
checkFormula(formula, data)
}
\arguments{
\item{formula}{formula of variables to check}

\item{data}{data.frame storing variables in the formula}
}
\description{
Check that variables in formula are present in the data
}
