% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crumblr.R
\name{clr}
\alias{clr}
\title{Centered log ratio transform}
\usage{
clr(counts, pseudocount = 0.5)
}
\arguments{
\item{counts}{count data with samples as rows and variables are columns}

\item{pseudocount}{added to counts to avoid issues with zeros}
}
\description{
Compute the centered log ratio (clr) transform of a count matrix.
}
\seealso{
compositions::clr
}
