% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamletProcessedData.R
\name{colData,dreamletProcessedData-method}
\alias{colData,dreamletProcessedData-method}
\title{Extract colData from \code{dreamletProcessedData}}
\usage{
\S4method{colData}{dreamletProcessedData}(x, ...)
}
\arguments{
\item{x}{A \code{dreamletProcessedData} object}

\item{...}{other arguments}
}
\value{
object from \code{colData} field
}
\description{
Extract colData from \code{dreamletProcessedData}
}
