% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamletProcessedData.R
\name{colData<-,dreamletProcessedData,ANY-method}
\alias{colData<-,dreamletProcessedData,ANY-method}
\title{Set colData}
\usage{
\S4method{colData}{dreamletProcessedData,ANY}(x, ...) <- value
}
\arguments{
\item{x}{\code{dreamletProcessedData} object}

\item{...}{other arguments}

\item{value}{\code{data.frame} or object that can be coerced to it}
}
\value{
none
}
\description{
Set colData of dreamletProcessedData, and check for same dimensions and rownames
}
