% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilr_composition_tests.R
\name{compositionTest}
\alias{compositionTest}
\title{Regression testing for composition changes}
\usage{
compositionTest(
  counts,
  formula,
  info,
  pseudocount = 0.1,
  useWeights = TRUE,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{counts}{count data}

\item{formula}{regression formula for independent variables}

\item{info}{data.frame storing variables in formula}

\item{pseudocount}{added to counts to avoid issues with zeros}

\item{useWeights}{default: TRUE. If TRUE use precision weights, else ignore weights}

\item{BPPARAM}{parameters for parallel evaluation}
}
\description{
Regression testing for composition changes using linear (mixed) models using log ratio of counts
}
