% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crumblr.R
\name{crumblr}
\alias{crumblr}
\title{Count ratio uncertainty modeling base linear regression}
\usage{
crumblr(counts, pseudocount = 0.5)
}
\arguments{
\item{counts}{count data with samples as rows and variables are columns}

\item{pseudocount}{added to counts to avoid issues with zeros}
}
\value{
An \code{EList} object with the following components:
\itemize{
 \item{E:}{numeric matrix of CLR transformed counts}
 \item{weights:}{numeric matrix of observation-level inverse-variance weights}
}
}
\description{
\deqn{\text{clr}_i({\bf \hat p}) = \log(p_i) - \frac{1}{D}\sum_{j=1}^D \log(p_j)}
 with sampling variance
\deqn{\text{var}[\text{clr}_i({\bf \hat p})] =  \frac{1}{n} \left[ \frac{1}{p_i} - \frac{2}{ D p_i} + \frac{1}{D^2}\sum_{j=1}^D p_j  \right] }
}
\details{
Count ratio uncertainty modeling base linear regression (crumblr) returns transformed counts and
}
