% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamletPairs.R
\name{dreamletPairs}
\alias{dreamletPairs}
\title{Differential expression between pair of assays}
\usage{
dreamletPairs(
  pb,
  assay.pairs,
  method = c("random", "fixed", "none"),
  formula = ~1,
  min.cells = 10,
  useCountsWeights = TRUE,
  isCounts = TRUE,
  normalize.method = "TMM",
  min.count = 10,
  robust = FALSE,
  quiet = FALSE,
  BPPARAM = SerialParam(),
  ...
)
}
\arguments{
\item{pb}{pseudobulk data as SingleCellExperiment object}

\item{assay.pairs}{array of two entries specifying assays (i.e. cell clusters) to compare}

\item{method}{account for repeated measures from donors using a "random" effect, a "fixed" effect, or "none"}

\item{formula}{covariates to include in the analysis.}

\item{min.cells}{minimum number of observed cells for a sample to be included in the analysis}

\item{useCountsWeights}{use cell count weights}

\item{isCounts}{logical, indicating if data is raw counts}

\item{normalize.method}{normalization method to be used by \code{calcNormFactors}}

\item{min.count}{minimum number of reads for a gene to be consider expressed in a sample.  Passed to \code{edgeR::filterByExpr}}

\item{robust}{logical, use eBayes method that is robust to outlier genes}

\item{quiet}{show messages}

\item{BPPARAM}{parameters for parallel evaluation}

\item{...}{other arguments passed to \code{dream}}
}
\description{
Perform differential expression between a pair of assays using linear (mixed) models
}
\examples{

library(muscat)
library(SingleCellExperiment)

data(example_sce)

# create pseudobulk for each sample and cell cluster
pb <- aggregateToPseudoBulk(example_sce, 
	assay = "counts",    
	cluster_id = 'cluster_id', 
	sample_id = 'sample_id',
	verbose=FALSE)

# compare first two assays (i.e. cell types)
ct.pairs = assayNames(pb)[1:2]

# run comparison
# use method = 'fixed' here since it is faster
res = dreamletPairs( pb, ct.pairs, method="fixed")

head(res)

}
