% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamlet.R
\name{[,dreamletProcessedData,ANY,ANY,ANY-method}
\alias{[,dreamletProcessedData,ANY,ANY,ANY-method}
\alias{[,dreamletProcessedData,dreamletProcessedData-method}
\title{Subset with brackets}
\usage{
\S4method{[}{dreamletProcessedData,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{x}{dreamletProcessedData object}

\item{i}{indeces to extract}
}
\description{
Subset with brackets
}
