% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamletProcessedData.R
\name{extractData}
\alias{extractData}
\alias{extractData,dreamletProcessedData,character-method}
\alias{extractData,dreamletProcessedData-method}
\title{Extract expression and colData}
\usage{
extractData(x, assay)

\S4method{extractData}{dreamletProcessedData,character}(x, assay)
}
\arguments{
\item{x}{A \code{dreamletProcessedData} object}

\item{assay}{assay to extract}
}
\description{
Extract expression and colData \code{dreamletProcessedData}
}
\details{
\code{DataFrame} of merge exrpession and colData
}
