% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamlet.R
\name{fitVarPart}
\alias{fitVarPart}
\title{Variance Partition analysis for each assay}
\usage{
fitVarPart(x, formula, data, BPPARAM = bpparam(), ...)
}
\arguments{
\item{x}{SingleCellExperiment or dreamletProcessedData object}

\item{formula}{regression formula for differential expression analysis}

\item{BPPARAM}{parameters for parallel evaluation}

\item{...}{other arguments passed to \code{dream}}
}
\description{
Perform Variance Partition analysis  for each assay
}
