% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilr_composition_tests.R
\name{ilrWithPrecisionWeights}
\alias{ilrWithPrecisionWeights}
\title{Apply isometric log ratio (ilr) transform and compute precision weights}
\usage{
ilrWithPrecisionWeights(counts, pseudocount = 0.1)
}
\arguments{
\item{counts}{data matrix of counts}

\item{pseudocount}{pseudocounts to avoid problems with zero counts}
}
\description{
Compute isometric log ratio (ilr) transform of counts, and their observation-level variances using the asymptotic normal approximation of Egozcue, et al (2020).  Convert observation-level variances to precisions in \code{Elist} for downstream use in \code{dream()}.
}
\details{
Apply isometric log ratio (ilr) transform and compute precision weights
}
\references{
{
  \insertRef{egozcue2020some}{dreamlet}
}
}
