% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamletProcessedData.R
\name{metadata,dreamletProcessedData-method}
\alias{metadata,dreamletProcessedData-method}
\title{Extract metadata from dreamletProcessedData}
\usage{
\S4method{metadata}{dreamletProcessedData}(x)
}
\arguments{
\item{x}{A dreamletProcessedData object}
}
\description{
Extract metadata from dreamletProcessedData
}
