% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTypeCompositionTest.R
\name{plotCellComposition}
\alias{plotCellComposition}
\title{Bar plot of cell compositions}
\usage{
plotCellComposition(obj, col, width = NULL)
}
\arguments{
\item{obj}{\code{SingleCellExperiment} from \code{aggregateToPseudoBulk}}

\item{col}{array of colors.  If missing, use default colors.  If \code{names(col)} is the same as \code{arrayNames(obj)}, then colors will be assigned by assay name#'}

\item{width}{specify width of bars}
}
\description{
Bar plot of cell compositions
}
