% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotForrest.R
\name{plotForrest}
\alias{plotForrest}
\title{Forrest plot}
\usage{
plotForrest(x, coef, gene)
}
\arguments{
\item{x}{result from \code{dreamlet}}

\item{coef}{coefficient to test with \code{topTable}}

\item{gene}{gene to show results for}
}
\description{
Forrest plot
}
