% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\docType{methods}
\name{plotHeatmap}
\alias{plotHeatmap}
\alias{plotHeatmap,cellSpecificityValues-method}
\alias{plotHeatmap,cellSpecificityValues,cellSpecificityValues-method}
\alias{plotHeatmap,matrix-method}
\alias{plotHeatmap,matrix,matrix-method}
\title{Plot heatmap}
\usage{
plotHeatmap(x, genes = rownames(x), color = "darkblue")

\S4method{plotHeatmap}{cellSpecificityValues}(x, genes = rownames(x), color = "darkblue")

\S4method{plotHeatmap}{matrix}(x, genes = rownames(x), color = "darkblue")
}
\arguments{
\item{x}{fractions for each gene}

\item{genes}{name of genes to plot}

\item{color}{color of heatmap}
}
\value{
heatmap
}
\description{
Plot heatmap
}
