% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R
\name{plotPCA}
\alias{plotPCA}
\alias{plotPCA,dreamletProcessedData-method}
\title{Plot PCA of gene expression for an assay}
\usage{
\S4method{plotPCA}{dreamletProcessedData}(
  object,
  assays = assayNames(object),
  nPC = 2,
  maxOutlierZ = 20,
  nrow = 2,
  size = 1
)
}
\arguments{
\item{object}{\code{dreamletProcessedData} from \code{processAssays()}}

\item{assays}{assays / cell types to analyze}

\item{nPC}{number of PCs to uses for outlier score with \code{outlier()}}

\item{maxOutlierZ}{cap outlier z-scores at this value for plotting to maintain consistent color scale}

\item{nrow}{number of rows in plot}

\item{size}{size passed to \code{geom_point()}}
}
\description{
Compute PCA of gene expression for an assay, and plot samples coloring by outlier score
}
\examples{
library(muscat)
library(SingleCellExperiment)

data(example_sce)

# create pseudobulk for each sample and cell cluster
pb <- aggregateToPseudoBulk(example_sce,
  assay = "counts",
  cluster_id = "cluster_id",
  sample_id = "sample_id",
  verbose = FALSE
)

# voom-style normalization
res.proc <- processAssays(pb, ~group_id)

# PCA to identify outliers
plotPCA( res.proc, "CD14+ Monocytes")
}
