% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpDF.R
\name{plotPercentBars,vpDF-method}
\alias{plotPercentBars,vpDF-method}
\alias{plotPercentBars,vpDF,vpDF-method}
\title{Bar plot of variance fractions}
\usage{
\S4method{plotPercentBars}{vpDF}(
  varPart,
  col = c(ggColorHue(ncol(varPart) - 3), "grey85"),
  width = NULL
)
}
\arguments{
\item{varPart}{object returned by extractVarPart() or fitExtractVarPartModel()}

\item{col}{color of bars for each variable}

\item{width}{specify width of bars}
}
\value{
Bar plot showing variance fractions for each gene
}
\description{
Bar plot of variance fractions for a subset of genes
}
\examples{
 
library(muscat)
library(SingleCellExperiment)

data(example_sce)

# create pseudobulk for each sample and cell cluster
pb <- aggregateToPseudoBulk(example_sce, 
   assay = "counts",    
   cluster_id = 'cluster_id', 
   sample_id = 'sample_id',
   verbose=FALSE)

# voom-style normalization
res.proc = processAssays( pb, ~ group_id)

# variance partitioning analysis
vp = fitVarPart( res.proc, ~ group_id)

# Show variance fractions at the gene-level for each cell type
genes = vp$gene[2:4]
plotPercentBars(vp[vp$gene \%in\% genes,])

}
