% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPercentBars.R, R/plotVarPart.R
\name{plotPercentBars,vpDF-method}
\alias{plotPercentBars,vpDF-method}
\alias{plotVarPart,vpDF,vpDF-method}
\alias{plotVarPart,DataFrame-method}
\alias{plotVarPart,DataFrame,DataFrame-method}
\title{Bar plot of variance fractions}
\usage{
\S4method{plotPercentBars}{vpDF}(
  varPart,
  col = c(ggColorHue(ncol(varPart) - 3), "grey85"),
  width = NULL
)

\S4method{plotVarPart}{DataFrame}(
  obj,
  col = c(ggColorHue(base::ncol(obj) - 3), "grey85"),
  label.angle = 20,
  main = "",
  ylab = "",
  convertToPercent = TRUE,
  ...
)
}
\arguments{
\item{varPart}{object returned by extractVarPart() or fitExtractVarPartModel()}

\item{col}{vector of colors}

\item{width}{specify width of bars}

\item{obj}{\code{varParFrac} object returned by \code{fitExtractVarPart} or \code{extractVarPart}}

\item{label.angle}{angle of labels on x-axis}

\item{main}{title of plot}

\item{ylab}{text on y-axis}

\item{convertToPercent}{multiply fractions by 100 to convert to percent values}

\item{...}{additional arguments}
}
\description{
Bar plot of variance fractions for a subset of genes

Violin plot of variance fraction for each gene and each variable
}
