% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotViolin.R
\docType{methods}
\name{plotViolin}
\alias{plotViolin}
\alias{plotViolin,cellSpecificityValues-method}
\alias{plotViolin,cellSpecificityValues,cellSpecificityValues-method}
\title{Plot Violins}
\usage{
plotViolin(x)

\S4method{plotViolin}{cellSpecificityValues}(x)
}
\arguments{
\item{x}{fractions for each gene}
}
\value{
Violin plot
}
\description{
Plot Violins
}
