% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVolcano.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Volcano plot}
\usage{
plotVolcano(
  df,
  showGenes = NULL,
  size = 15,
  minp = 9.99999999999997e-311,
  cutoff = 0.05,
  xlim = NULL,
  ymax = NULL
)
}
\arguments{
\item{df}{data.frame produced by \code{topTable}}

\item{showGenes}{array of gene names corresponding rownames of \code{df} to highlight}

\item{size}{text size}

\item{minp}{minimum p-value to show on the y-axis}

\item{cutoff}{adj.P.Val cutoff to distinguish significant from non-significant genes}

\item{xlim}{limits on the x-axis}

\item{ymax}{maximum value on the y-axis}
}
\description{
Volcano plot
}
