% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVolcano.R
\name{plotVolcano_grid}
\alias{plotVolcano_grid}
\title{Volcano plot for each cell type}
\usage{
plotVolcano_grid(
  x,
  coef,
  nGenes = 5,
  size = 12,
  minp = 9.99999999999997e-311,
  cutoff = 0.05,
  ncol = 3
)
}
\arguments{
\item{x}{result from \code{dreamlet}}

\item{coef}{coefficient to test with \code{topTable}}

\item{nGenes}{number of genes to highlight in each volcano plot}

\item{size}{text size}

\item{minp}{minimum p-value to show on the y-axis}

\item{cutoff}{adj.P.Val cutoff to distinguish significant from non-significant genes}

\item{ncol}{number of columns in the plot}
}
\description{
Volcano plot for each cell type
}
