% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVoom.R
\docType{methods}
\name{plotVoom}
\alias{plotVoom}
\alias{plotVoom,dreamletProcessedData-method}
\alias{plotVoom,dreamletProcessedData,dreamletProcessedData-method}
\alias{plotVoom,list-method}
\alias{plotVoom,list,list-method}
\title{Plot voom curves from each cell type}
\usage{
plotVoom(x, ncol = 3)

\S4method{plotVoom}{dreamletProcessedData}(x, ncol = 3)

\S4method{plotVoom}{list}(x, ncol = 3)
}
\arguments{
\item{x}{dreamletProcessedData}

\item{ncol}{number of columns in the plot}
}
\description{
Plot voom curves from each cell type
}
