% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVoom.R
\name{plotVoom}
\alias{plotVoom}
\title{Plot voom curves from each cell type}
\usage{
plotVoom(x, ncol = 3)
}
\arguments{
\item{x}{dreamletProcessedData}

\item{ncol}{number of columns in the plot}
}
\description{
Plot voom curves from each cell type
}
