% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotZenithResults.R
\name{plotZenithResults}
\alias{plotZenithResults}
\title{Heatmap of zenith results}
\usage{
plotZenithResults(df, ntop = 5, nbottom = 5)
}
\arguments{
\item{df}{result \code{data.frame} from \link{zenith_gsa}}

\item{ntop}{number of gene sets with highest t-statistic to show}

\item{nbottom}{number of gene sets with lowest t-statistic to show}
}
\description{
Heatmap of zenith results showing genesets that have the top and bottom t-statistics from each assay.
}
