% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamlet.R, R/dreamletProcessedData.R
\name{print,dreamletResult-method}
\alias{print,dreamletResult-method}
\alias{print,dreamletResult,dreamletResult-method}
\alias{print,dreamletProcessedData-method}
\alias{print,dreamletProcessedData,dreamletProcessedData-method}
\title{Print object}
\usage{
\S4method{print}{dreamletResult}(x, ...)

\S4method{print}{dreamletProcessedData}(x, ...)
}
\arguments{
\item{x}{\code{dreamletProcessedData} object}

\item{...}{other arguments}
}
\value{
print data stored in object
}
\description{
Print object

Print object
}
