% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{residuals,dreamletResult-method}
\alias{residuals,dreamletResult-method}
\alias{residuals,dreamletResult,dreamletResult-method}
\title{Extract residuls from \code{dreamletResult}}
\usage{
\S4method{residuals}{dreamletResult}(object, ...)
}
\arguments{
\item{object}{\code{dreamletResult} object}

\item{...}{other arguments}
}
\value{
residuals from model fit
}
\description{
Extract residuls from \code{dreamletResult}
}
\examples{
library(muscat)
library(SingleCellExperiment)

data(example_sce)

# create pseudobulk for each sample and cell cluster
pb <- aggregateToPseudoBulk(example_sce, 
   assay = "counts",    
   cluster_id = 'cluster_id', 
   sample_id = 'sample_id',
   verbose=FALSE)

# voom-style normalization
res.proc = processAssays( pb, ~ group_id)

# Differential expression analysis within each assay,
# evaluated on the voom normalized data 
res.dl = dreamlet( res.proc, ~ group_id)

# extract residuals for each assay (i.e. cell type)
# Return list with entry for each assay with for retained samples and genes
resid.lst = residuals(res.dl)

}
