% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamlet.R
\name{show,dreamletProcessedData-method}
\alias{show,dreamletProcessedData-method}
\alias{show,dreamletProcessedData,dreamletProcessedData-method}
\title{Show object}
\usage{
\S4method{show}{dreamletProcessedData}(object)
}
\arguments{
\item{object}{dreamletProcessedData object}
}
\description{
Show object
}
