% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpDF.R
\name{sortCols,vpDF-method}
\alias{sortCols,vpDF-method}
\title{Sort variance partition statistics}
\usage{
\S4method{sortCols}{vpDF}(
  x,
  FUN = median,
  decreasing = TRUE,
  last = c("Residuals", "Measurement.error"),
  ...
)
}
\arguments{
\item{x}{object returned by \code{extractVarPart()} or \code{fitExtractVarPartModel()}}

\item{FUN}{function giving summary statistic to sort by.  Defaults to median}

\item{decreasing}{logical.  Should the sorting be increasing or decreasing?}

\item{last}{columns to be placed on the right, regardless of values in these columns}

\item{...}{other arguments to sort}
}
\description{
Sort variance partition statistics
}
