% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamletProcessedData.R
\name{subsetSamples}
\alias{subsetSamples}
\title{Extract a subset of samples}
\usage{
subsetSamples(x, ids)
}
\arguments{
\item{x}{dreamletProcessedData}

\item{ids}{column names to retain}
}
\description{
Extract a subset of samples
}
