% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash.R
\name{tabToMatrix}
\alias{tabToMatrix}
\title{Convert results table to matrix}
\usage{
tabToMatrix(tab, col, rn = "ID", cn = "assay")
}
\arguments{
\item{tab}{results table from \code{topTable()}}

\item{col}{which column to extract}

\item{rn}{column id storing rownames}

\item{cn}{column id storing colnames}
}
\description{
Convert results table to matrix
}
