% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilr_composition_tests.R
\name{topTable,MArrayLM2_ilr-method}
\alias{topTable,MArrayLM2_ilr-method}
\alias{toptable,MArrayLM2_ilr-method}
\title{toptable for MArrayLM2_ilr}
\usage{
\S4method{topTable}{MArrayLM2_ilr}(
  fit,
  coef = NULL,
  number = 10,
  genelist = fit$genes,
  adjust.method = "BH",
  sort.by = "B",
  resort.by = NULL,
  p.value = 1,
  lfc = 0,
  confint = FALSE
)
}
\arguments{
\item{fit}{fit}

\item{coef}{coef}

\item{number}{number}

\item{genelist}{genelist}

\item{adjust.method}{adjust.method}

\item{sort.by}{sort.by}

\item{resort.by}{resort.by}

\item{p.value}{p.value}

\item{lfc}{lfc}

\item{confint}{confint}
}
\value{
results of toptable
}
\description{
toptable for MArrayLM2_ilr
}
