% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamlet.R
\name{topTable,dreamletResult-method}
\alias{topTable,dreamletResult-method}
\alias{topTable,dreamletResult,dreamletResult-method}
\title{Table of Top Genes from dreamlet fit}
\usage{
\S4method{topTable}{dreamletResult}(
  fit,
  coef = NULL,
  number = 10,
  genelist = NULL,
  adjust.method = "BH",
  sort.by = "P",
  resort.by = NULL,
  p.value = 1,
  lfc = 0,
  confint = FALSE
)
}
\arguments{
\item{fit}{dreamletResult object}

\item{coef}{coef}

\item{number}{number}

\item{genelist}{genelist}

\item{adjust.method}{adjust.method}

\item{sort.by}{sort.by}

\item{resort.by}{resort.by}

\item{p.value}{p.value}

\item{lfc}{lfc}

\item{confint}{confint}
}
\description{
Extract a table of the top-ranked genes from a dreamlet fit.
}
