% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpDF.R
\docType{class}
\name{vpDF-class}
\alias{vpDF-class}
\alias{assayNames,vpDF-method}
\alias{assay,vpDF,ANY-method}
\title{Class vpDF}
\usage{
\S4method{assayNames}{vpDF}(x, ...)

\S4method{assay}{vpDF,ANY}(x, i, withDimnames = TRUE, ...)
}
\arguments{
\item{x}{vpDF object}

\item{...}{additional arguments}

\item{i}{number indicating index, or string indicating assay}

\item{withDimnames}{not used}
}
\description{
Class \code{vpDF} stores results for each gene for each assay

Get aassayNames

Get assays by name
}
