% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenith_gsa.R
\name{zenith_gsa}
\alias{zenith_gsa}
\alias{zenith_gsa,dreamletResult,ANY,GeneSetCollection-method}
\title{Perform zenith analysis}
\usage{
zenith_gsa(
  fit,
  coefs,
  geneSets,
  n_genes_min = 10,
  inter.gene.cor = 0.01,
  progressbar = TRUE
)

\S4method{zenith_gsa}{dreamletResult,ANY,GeneSetCollection}(
  fit,
  coefs,
  geneSets,
  n_genes_min = 10,
  inter.gene.cor = 0.01,
  progressbar = TRUE
)
}
\arguments{
\item{fit}{results from \code{dreamlet}}

\item{coefs}{coefficients to test using \code{topTable(fit, coef)}}

\item{geneSets}{a list of gene sets}

\item{n_genes_min}{minumum number of genes in a geneset}

\item{inter.gene.cor}{if NA, estimate correlation from data.  Otherwise, use specified value}

\item{progressbar}{if TRUE, show progress bar}
}
\value{
\code{data.frame} of results for each gene set and cell type
}
\description{
Perform zenith analysis
}
\examples{
library(muscat)
library(SingleCellExperiment)

data(example_sce)

# create pseudobulk for each sample and cell cluster
pb <- aggregateToPseudoBulk(example_sce, 
   assay = "counts",    
   cluster_id = 'cluster_id', 
   sample_id = 'sample_id',
   verbose=FALSE)

# voom-style normalization
res.proc = processAssays( pb, ~ group_id)

# Differential expression analysis within each assay,
# evaluated on the voom normalized data 
res.dl = dreamlet( res.proc, ~ group_id)

# Load Gene Ontology database 
# use gene 'SYMBOL', or 'ENSEMBL' id
# use get_MSigDB() to load MSigDB
library(zenith)
go.gs = get_GeneOntology("CC", to="SYMBOL")
   
# Run zenith gene set analysis on result of dreamlet
res_zenith = zenith_gsa(res.dl, coef = 'group_idstim', go.gs)

# for each cell type select 5 genesets with largest t-statistic
# and 1 geneset with the lowest
# Grey boxes indicate the gene set could not be evaluted because
#    to few genes were represented
plotZenithResults(res_zenith, 5, 1)

}
