% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateNonCountSignal.R
\name{aggregateNonCountSignal}
\alias{aggregateNonCountSignal}
\title{Aggregation of single-cell signals}
\usage{
aggregateNonCountSignal(
  sce,
  assay = NULL,
  sample_id = NULL,
  cluster_id = NULL,
  verbose = TRUE,
  BPPARAM = SerialParam(progressbar = verbose)
)
}
\arguments{
\item{sce}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{assay}{character string specifying the assay slot to use as 
input data. Defaults to the 1st available (\code{assayNames(x)[1]}).}

\item{sample_id}{character string specifying which variable to use as sample id}

\item{cluster_id}{character string specifying which variable to use as cluster id}

\item{verbose}{logical. Should information on progress be reported?}

\item{BPPARAM}{a \code{\link[BiocParallel]{BiocParallelParam}}
object specifying how aggregation should be parallelized.}
}
\value{
a \code{dreamletProcessedData} object
}
\description{
Aggregation of single-cell to pseudobulk data for non-count data.
}
\details{
For standard analysis of count data, \code{aggregateToPseudoBulk()} create pseudobulk counts and \code{processAssays()} performs folder and estimates precision weights.  Yet the \code{dreamlet} workflow can also be applied to non-count data. In this case, a signal is averaged across all cells from a given sample and cell type.  And the precision weights step is set skipped.
}
