% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateToPseudoBulk.R
\name{aggregateToPseudoBulk}
\alias{aggregateToPseudoBulk}
\title{Aggregation of single-cell to pseudobulk data}
\usage{
aggregateToPseudoBulk(
  x,
  assay = NULL,
  sample_id = NULL,
  cluster_id = NULL,
  fun = c("sum", "mean", "median", "prop.detected", "num.detected"),
  scale = FALSE,
  verbose = TRUE,
  BPPARAM = SerialParam(progressbar = verbose)
)
}
\arguments{
\item{x}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{assay}{character string specifying the assay slot to use as 
input data. Defaults to the 1st available (\code{assayNames(x)[1]}).}

\item{sample_id}{character string specifying which variable to use as sample id}

\item{cluster_id}{character string specifying which variable to use as cluster id}

\item{fun}{a character string.
Specifies the function to use as summary statistic.
Passed to \code{summarizeAssayByGroup2}.}

\item{scale}{logical. Should pseudo-bulks be scaled
with the effective library size & multiplied by 1M?}

\item{verbose}{logical. Should information on progress be reported?}

\item{BPPARAM}{a \code{\link[BiocParallel]{BiocParallelParam}}
object specifying how aggregation should be parallelized.}
}
\value{
a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.

 Aggregation parameters (\code{assay, by, fun, scaled}) are stored in  \code{metadata()$agg_pars}, where \code{by = c(cluster_id, sample_id)}.  The number of cells that were aggregated are accessible in \code{int_colData()$n_cells}.
}
\description{
Aggregation of single-cell to pseudobulk data.  Adapted from \code{muscat::aggregateData} and has same syntax and results.  But can be much faster in for \code{SingleCellExperiment} backed by H5AD files
}
\details{
Adapted from \code{muscat::aggregateData} and has simular syntax and same results.  But can be much faster in for \code{SingleCellExperiment} backed by H5AD files because this summarizes counts using \code{\link[DelayedMatrixStats]{DelayedMatrixStats}}.
}
\examples{

library(muscat)
library(SingleCellExperiment)

data(example_sce)

# create pseudobulk for each sample and cell cluster
pb <- aggregateToPseudoBulk(example_sce, 
   assay = "counts",    
   cluster_id = 'cluster_id', 
   sample_id = 'sample_id',
   verbose=FALSE)

# pseudobulk data from each cell type
# is stored as its own assay 
pb

# aggregate by cluster only, 
# collapsing all samples into the same pseudobulk
pb2 <- aggregateToPseudoBulk(example_sce, cluster_id = "cluster_id", verbose=FALSE)

pb2  

}
\references{
Crowell, HL, Soneson, C, Germain, P-L, Calini, D, 
Collin, L, Raposo, C, Malhotra, D & Robinson, MD: 
On the discovery of population-specific state transitions from 
multi-sample multi-condition single-cell RNA sequencing data. 
\emph{bioRxiv} \strong{713412} (2018). 
doi: \url{https://doi.org/10.1101/713412}
}
\author{
Gabriel Hoffman, Helena L Crowell & Mark D Robinson
}
