% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamlet.R
\name{assay,dreamletProcessedData,ANY-method}
\alias{assay,dreamletProcessedData,ANY-method}
\alias{assay,dreamletProcessedData,dreamletProcessedData-method}
\title{Get assay}
\usage{
\S4method{assay}{dreamletProcessedData,ANY}(x, i, withDimnames = TRUE, ...)
}
\arguments{
\item{x}{dreamletProcessedData object}

\item{i}{number indicating index, or string indicating assay}

\item{withDimnames}{not used}

\item{...}{other arguments}
}
\description{
Get assay
}
