% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamlet.R, R/dreamletProcessedData.R, R/vpDF.R
\name{assayNames,dreamletResult-method}
\alias{assayNames,dreamletResult-method}
\alias{assayNames,dreamletResult,dreamletResult-method}
\alias{assayNames,dreamletProcessedData-method}
\alias{assayNames,dreamletProcessedData,dreamletProcessedData-method}
\alias{assayNames,vpDF-method}
\alias{assayNames,vpDF,vpDF-method}
\title{Get assayNames}
\usage{
\S4method{assayNames}{dreamletResult}(x, ...)

\S4method{assayNames}{dreamletProcessedData}(x, ...)

\S4method{assayNames}{vpDF}(x, ...)
}
\arguments{
\item{x}{vpDF object}

\item{...}{additional arguments}
}
\value{
array of assay names
}
\description{
Get assayNames

Get assayNames

Get assayNames
}
