% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamlet.R
\name{assayNames,dreamletProcessedData-method}
\alias{assayNames,dreamletProcessedData-method}
\alias{assayNames,dreamletProcessedData,dreamletProcessedData-method}
\title{Get assayNames}
\usage{
\S4method{assayNames}{dreamletProcessedData}(x, ...)
}
\arguments{
\item{x}{dreamletProcessedData object}

\item{...}{other arguments}
}
\description{
Get assayNames
}
