% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTypeSpecificity.R
\docType{class}
\name{cellSpecificityValues-class}
\alias{cellSpecificityValues-class}
\title{Class cellSpecificityValues}
\description{
Class \code{cellSpecificityValues} cell type specificity values for each gene and cell type
}
