% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTypeSpecificity.R
\name{cellTypeSpecificity}
\alias{cellTypeSpecificity}
\title{Get cell type specificity of gene expression}
\usage{
cellTypeSpecificity(pb, ...)
}
\arguments{
\item{pb}{\code{SingleCellExperiment} of pseudobulk data where easy \code{assay} is a cell type.}

\item{...}{other arguments passed to \code{edgeR::calcNormFactors()}}
}
\value{
matrix of the fraction of expression attributable to each cell type for each gene.
}
\description{
For each gene, compute fraction of overall expression attributable to each cell type
}
\details{
Sum counts for each cell type, and compute the fraction of counts-per-million attributable to each cell type for each gene
}
\examples{
library(muscat)
library(SingleCellExperiment)

data(example_sce)

# create pseudobulk for each sample and cell cluster
pb <- aggregateToPseudoBulk(example_sce, 
assay = "counts",    
cluster_id = 'cluster_id', 
sample_id = 'sample_id',
verbose=FALSE)

# Compute cell type specificity of each gene
df = cellTypeSpecificity( pb)

df_melt = reshape2::melt(df)

# Violin plot of cell type specificity
ggplot(df_melt, aes(Var2, value, fill=Var2)) + 
geom_violin() + 
geom_boxplot(width=.1, fill="grey") + 
theme_classic() + 
theme(aspect.ratio=1, legend.position="none", 
	plot.title = element_text(hjust = 0.5),
	axis.text.x = element_text(angle = 60, vjust = 1, hjust=1)) + 
scale_y_continuous(limits=c(0,1), expand=c(0,0)) + 
ylab("Fraction of gene expression") + 
 geom_hline(yintercept=1/ncol(df), color="grey60", linetype="dashed") +
ggtitle("Cell type specificity")

# Barplot of 5 genes
genes = rownames(df)[1:5]

ggplot(df_melt[df_melt$Var1 \%in\% genes,], aes(Var1, value, fill=Var2)) + 
	geom_bar(stat="identity") + 
	theme_classic() + 
	theme(aspect.ratio=1, plot.title = element_text(hjust = 0.5)) + 
	coord_flip() + 
	scale_fill_discrete(name = "Cell type") + 
	xlab("Gene") + 
	ylab("Fraction of gene expression") + 
	scale_y_continuous(limits=c(0, 1), expand=c(0,0)) + 
 ggtitle("Cell type specificity")

}
