% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamlet.R
\name{coefNames}
\alias{coefNames}
\alias{coefNames,dreamletResult-method}
\title{Get coefficient names}
\usage{
coefNames(obj)

\S4method{coefNames}{dreamletResult}(obj)
}
\arguments{
\item{obj}{A dreamletResult object}
}
\description{
Get coefficient names
}
