% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamletProcessedData.R
\name{colData,dreamletProcessedData-method}
\alias{colData,dreamletProcessedData-method}
\title{Extract colData from dreamletProcessedData}
\usage{
\S4method{colData}{dreamletProcessedData}(x, ...)
}
\arguments{
\item{x}{A dreamletProcessedData object}

\item{...}{other arguments}
}
\description{
Extract colData from dreamletProcessedData
}
