% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLogCPM.R
\name{computeLogCPM}
\alias{computeLogCPM}
\title{Compute log normalized counts}
\usage{
computeLogCPM(sce, prior.count = 2)
}
\arguments{
\item{sce}{\code{SingleCellExperiment} with counts stored as \code{counts(sce)}}

\item{prior.count}{average count to be added to each observation to avoid taking log of zero}
}
\value{
matrix of log CPM values
}
\description{
Compute normalized counts as log2 counts per million
}
\details{
This function gives same result as \code{edgeR::cpm(counts(sce), log=TRUE)}
}
\examples{
library(muscat)
library(SingleCellExperiment)

data(example_sce)

logcounts(example_sce) = computeLogCPM(example_sce)
}
\seealso{
also \code{edgeR::cpm()}
}
