% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamletProcessedData.R, R/vpDF.R
\name{details}
\alias{details}
\alias{details,dreamletProcessedData-method}
\alias{details,dreamletResult-method}
\alias{details,vpDF-method}
\title{Extract details from dreamletProcessedData}
\usage{
details(object)

\S4method{details}{dreamletProcessedData}(object)

\S4method{details}{dreamletResult}(object)

\S4method{details}{vpDF}(object)
}
\arguments{
\item{object}{A dreamletProcessedData object}
}
\description{
Extract details from dreamletProcessedData
}
