% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamlet.R
\name{dreamlet}
\alias{dreamlet}
\title{Differential expression for each assay}
\usage{
dreamlet(
  x,
  formula,
  data,
  L,
  min.cells = 10,
  isCounts = TRUE,
  robust = FALSE,
  normalize.method = "TMM",
  BPPARAM = bpparam(),
  ...
)
}
\arguments{
\item{x}{SingleCellExperiment or dreamletProcessedData object}

\item{formula}{regression formula for differential expression analysis}

\item{L}{contrast matrix specifying a linear combination of fixed effects to test}

\item{min.cells}{minimum number of observed cells for a sample to be included in the analysis}

\item{isCounts}{logical, indicating if data is raw counts}

\item{robust}{logical, use eBayes method that is robust to outlier genes}

\item{normalize.method}{normalization method to be used by \code{calcNormFactors}}

\item{BPPARAM}{parameters for parallel evaluation}

\item{...}{other arguments passed to \code{dream}}
}
\description{
Perform differential expression for each assay using linear mixed models
}
