% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamlet.R
\name{dreamlet}
\alias{dreamlet}
\alias{dreamlet,dreamletProcessedData-method}
\title{Differential expression for each assay}
\usage{
dreamlet(
  x,
  formula,
  data = colData(x),
  L.list = NULL,
  min.cells = 10,
  isCounts = TRUE,
  robust = FALSE,
  normalize.method = "TMM",
  quiet = FALSE,
  BPPARAM = SerialParam(),
  ...
)

\S4method{dreamlet}{dreamletProcessedData}(
  x,
  formula,
  data = colData(x),
  L.list = NULL,
  min.cells = 10,
  isCounts = TRUE,
  robust = FALSE,
  normalize.method = "TMM",
  quiet = FALSE,
  BPPARAM = SerialParam(),
  ...
)
}
\arguments{
\item{x}{SingleCellExperiment or dreamletProcessedData object}

\item{formula}{regression formula for differential expression analysis}

\item{data}{metadata used in regression formula}

\item{L.list}{list of contrasts specifying linear combinations of fixed effects to tests}

\item{min.cells}{minimum number of observed cells for a sample to be included in the analysis}

\item{isCounts}{logical, indicating if data is raw counts}

\item{robust}{logical, use eBayes method that is robust to outlier genes}

\item{normalize.method}{normalization method to be used by \code{calcNormFactors}}

\item{quiet}{show messages}

\item{BPPARAM}{parameters for parallel evaluation}

\item{...}{other arguments passed to \code{dream}}
}
\description{
Perform differential expression for each assay using linear mixed models
}
