% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamletCompareClusters.R
\name{dreamletCompareClusters}
\alias{dreamletCompareClusters}
\title{Differential expression between pair of assays}
\usage{
dreamletCompareClusters(
  pb,
  assays,
  method = c("random", "fixed", "none"),
  formula = ~1,
  min.cells = 10,
  useCountsWeights = TRUE,
  isCounts = TRUE,
  normalize.method = "TMM",
  min.count = 10,
  robust = FALSE,
  quiet = FALSE,
  BPPARAM = SerialParam(),
  ...
)
}
\arguments{
\item{pb}{pseudobulk data as SingleCellExperiment object}

\item{assays}{array of two entries specifying assays (i.e. cell clusters) to compare, of a list of two sets of assays.}

\item{method}{account for repeated measures from donors using a "random" effect, a "fixed" effect, or "none"}

\item{formula}{covariates to include in the analysis.}

\item{min.cells}{minimum number of observed cells for a sample to be included in the analysis}

\item{useCountsWeights}{use cell count weights}

\item{isCounts}{logical, indicating if data is raw counts}

\item{normalize.method}{normalization method to be used by \code{calcNormFactors}}

\item{min.count}{minimum number of reads for a gene to be consider expressed in a sample.  Passed to \code{edgeR::filterByExpr}}

\item{robust}{logical, use eBayes method that is robust to outlier genes}

\item{quiet}{show messages}

\item{BPPARAM}{parameters for parallel evaluation}

\item{...}{other arguments passed to \code{dream}}
}
\description{
Perform differential expression between a pair of assays using linear (mixed) models
}
\examples{

library(muscat)
library(SingleCellExperiment)

data(example_sce)

# create pseudobulk for each sample and cell cluster
pb <- aggregateToPseudoBulk(example_sce, 
	assay = "counts",    
	cluster_id = 'cluster_id', 
	sample_id = 'sample_id',
	verbose=FALSE)

# compare first two assays (i.e. cell types)
ct.pairs =  c("B cells", "CD14+ Monocytes")

# run comparison
# use method = 'fixed' here since it is faster
res = dreamletCompareClusters( pb, ct.pairs, method="fixed")

# genes with highest logFC are most highly expressed in 
# B cells compared to CD14+ Monocytes
head(res)

# compare B cells versus the rest of the cell types
# 'rest' is a keyword indicator all other assays	
res = dreamletCompareClusters( pb, c("B cells", 'rest'), method="fixed")

# genes with highest logFC are most highly expressed in 
# B cells compared to all others
head(res)

lst = list( test = c("CD14+ Monocytes", "FCGR3A+ Monocytes"), 
		baseline = c("CD4 T cells", "CD8 T cells"))

# compare 2 monocyte clusters to two T cell clusters
res = dreamletCompareClusters( pb, lst, method="fixed")

# genes with highest logFC are most highly expressed in 
# monocytes compared to T cells
head(res)

}
