% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffVar.R, R/dreamlet.R
\docType{class}
\name{dreamletResult-class}
\alias{dreamletResult-class}
\title{Class dreamletResult}
\value{
none

none
}
\description{
Class \code{dreamletResult} stores results produced by \code{dreamlet()} to give a standard interface for downstream analysis

Class \code{dreamletResult} stores results produced by \code{dreamlet()} to give a standard interface for downstream analysis
}
