% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{equalFormulas}
\alias{equalFormulas}
\title{Check if two formulas are equal}
\usage{
equalFormulas(formula1, formula2)
}
\arguments{
\item{formula1}{first formula}

\item{formula2}{second formula}
}
\description{
Check if two formulas are equal by evaluating the formulas and extracting terms
}
\examples{

# These formulas are equivalent
formula1 = ~ Size + 1
formula2 = ~ 1 + Size 

dreamlet:::equalFormulas( formula1, formula2)

}
