% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamlet.R, R/misc.R
\name{[,dreamletResult,ANY,ANY,ANY-method}
\alias{[,dreamletResult,ANY,ANY,ANY-method}
\alias{[,dreamletResult,dreamletResult-method}
\alias{topTable,dreamletResult-method}
\alias{topTable,dreamletResult,dreamletResult-method}
\alias{[,dreamletProcessedData,ANY,ANY,ANY-method}
\alias{[,dreamletProcessedData,dreamletProcessedData-method}
\title{Subset with brackets}
\usage{
\S4method{[}{dreamletResult,ANY,ANY,ANY}(x, i)

\S4method{topTable}{dreamletResult}(
  fit,
  coef = NULL,
  number = 10,
  genelist = NULL,
  adjust.method = "BH",
  sort.by = "P",
  resort.by = NULL,
  p.value = 1,
  lfc = 0,
  confint = FALSE
)

\S4method{[}{dreamletProcessedData,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{x}{dreamletProcessedData object}

\item{i}{indeces to extract}

\item{fit}{dreamletResult object}

\item{...}{arguments passed to topTable}
}
\description{
Subset with brackets

Extract a table of the top-ranked genes from a dreamlet fit.

Subset with brackets
}
