% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamlet.R, R/dreamletProcessedData.R
\name{[,dreamletResult,ANY,ANY,ANY-method}
\alias{[,dreamletResult,ANY,ANY,ANY-method}
\alias{[,dreamletResult,dreamletResult-method}
\alias{[,dreamletProcessedData,ANY,ANY,ANY-method}
\alias{[,dreamletProcessedData,dreamletProcessedData-method}
\title{Subset with brackets}
\usage{
\S4method{[}{dreamletResult,ANY,ANY,ANY}(x, i)

\S4method{[}{dreamletProcessedData,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{x}{\code{dreamletProcessedData} object}

\item{i}{indeces to extract}
}
\description{
Subset with brackets

Subset with brackets
}
