% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamletProcessedData.R
\name{extractData}
\alias{extractData}
\alias{extractData,dreamletProcessedData,character-method}
\alias{extractData,dreamletProcessedData-method}
\title{Extract expression and colData}
\usage{
extractData(x, assay)

\S4method{extractData}{dreamletProcessedData,character}(x, assay)
}
\arguments{
\item{x}{A \code{dreamletProcessedData} object}

\item{assay}{assay to extract}
}
\value{
\code{data.frame} or \code{DataFrame} of merged expression and colData
}
\description{
Extract expression and \code{colData} from \code{dreamletProcessedData}
}
\examples{
library(muscat)
library(SingleCellExperiment)

data(example_sce)

# create pseudobulk for each sample and cell cluster
pb <- aggregateToPseudoBulk(example_sce,
  assay = "counts",
  cluster_id = "cluster_id",
  sample_id = "sample_id",
  verbose = FALSE
)

# voom-style normalization
res.proc <- processAssays(pb, ~group_id)

# Extract data.frame of colData merged with expression.
# variables and genes are stored as columns, samples as rows
df_merge <- extractData(res.proc, "B cells")

dim(df_merge)

# first few columns
df_merge[, 1:6]

}
