% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitVarPart.R
\name{fitVarPart}
\alias{fitVarPart}
\alias{fitVarPart,dreamletProcessedData-method}
\title{Variance Partition analysis for each assay}
\usage{
fitVarPart(
  x,
  formula,
  data = colData(x),
  quiet = FALSE,
  BPPARAM = SerialParam(),
  ...
)

\S4method{fitVarPart}{dreamletProcessedData}(
  x,
  formula,
  data = colData(x),
  quiet = FALSE,
  BPPARAM = SerialParam(),
  ...
)
}
\arguments{
\item{x}{SingleCellExperiment or dreamletProcessedData object}

\item{formula}{regression formula for differential expression analysis}

\item{data}{metadata used in regression formula}

\item{quiet}{show messages}

\item{BPPARAM}{parameters for parallel evaluation}

\item{...}{other arguments passed to \code{dream}}
}
\description{
Perform Variance Partition analysis  for each assay
}
