% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{outlier}
\alias{outlier}
\title{Multivariate outlier detection}
\usage{
outlier(data, robust = TRUE, ...)
}
\arguments{
\item{data}{matrix of data}

\item{robust}{use robust methods}

\item{...}{arguments passed to \code{MASS::cov.rob()}}
}
\description{
Detect multivariante outliers using Mahalanobis distance using mean and covariance estimated either with standard or robust methods.
}
\details{
The distance follow a chisq distrubtion under the null with standard method for mean and covariance.  It is approximate of robust method is used.  So use \code{qchisq(p = 0.999 , df = 2)} to get cutoff to keep 99.9\% of samples under the null for data with 2 columns.
}
\examples{
data = matrix(rnorm(200), 100, 2)

res = outlier(data)

res[1:4]
}
