% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialPrecisionWeights.R
\name{pbWeights}
\alias{pbWeights}
\title{Compute observation weights for pseudobulk}
\usage{
pbWeights(
  sce,
  sample_id,
  cluster_id,
  shrink = TRUE,
  prior.count = 0.5,
  quantileOffset = 0.1,
  h5adBlockSizes = 1e+09
)
}
\arguments{
\item{sce}{\code{SingleCellExperiment} of where \code{counts(sce)} stores the raw count data at the single cell level}

\item{sample_id}{character string specifying which variable to use as sample id}

\item{cluster_id}{character string specifying which variable to use as cluster id}

\item{shrink}{Defaults to \code{TRUE}. Use empirical Bayes variance shrinkage from \code{limma} to shrink estimates of expression variance across cells within each sample}

\item{prior.count}{Defaults to \code{0.5}. Count added to each observation at the pseudobulk level.  This is scaled but the number of cells before added to the cell level}

\item{quantileOffset}{Defaults to \code{0.1}. When computing the precision from the variance, regularize the reciprocal by adding a small value to the denominator. For a gene with variances stored in the array \code{x}, add \code{quantile(x, quantileOffset)} before taking the reciprocal.}

\item{h5adBlockSizes}{set the automatic block size block size (in bytes) for DelayedArray to read an H5AD file.  Larger values use more memory but are faster.}
}
\description{
Compute observation weights for pseudobulk using the delta method to approximate the variance of the log2 counts per million considering variation in the number of cells and gene expression variance across cells within each sample.
}
\examples{
library(muscat)

data(example_sce)

# create pseudobulk for each sample and cell cluster
pb <- aggregateToPseudoBulk(example_sce,
  assay = "counts",
  cluster_id = "cluster_id",
  sample_id = "sample_id",
  verbose = FALSE
)

# Create precision weights for pseudobulk
weightsList <- pbWeights(example_sce, 
    cluster_id = "cluster_id",
    sample_id = "sample_id")

}
