% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialPrecisionWeights.R
\name{pbWeights}
\alias{pbWeights}
\title{Compute precision weights for pseudobulk}
\usage{
pbWeights(
  sce,
  sample_id,
  cluster_id,
  method = c("ncells", "delta"),
  shrink = TRUE,
  prior.count = 0.5,
  quantileOffset = 0.1,
  maxRatio = 20,
  h5adBlockSizes = 1e+09,
  verbose = TRUE
)
}
\arguments{
\item{sce}{\code{SingleCellExperiment} of where \code{counts(sce)} stores the raw count data at the single cell level}

\item{sample_id}{character string specifying which variable to use as sample id}

\item{cluster_id}{character string specifying which variable to use as cluster id}

\item{method}{select method to compute precision weights. \code{'ncells'} use the number of cells, this is faster. Subsequent arguments are ignored. \code{'delta'} use the delta method based on normal approximation to a negative binomial model, slower but can increase power.}

\item{shrink}{Defaults to \code{TRUE}. Use empirical Bayes variance shrinkage from \code{limma} to shrink estimates of expression variance across cells within each sample}

\item{prior.count}{Defaults to \code{0.5}. Count added to each observation at the pseudobulk level.  This is scaled but the number of cells before added to the cell level}

\item{quantileOffset}{Defaults to \code{0.1}. When computing the precision from the variance, regularize the reciprocal by adding a small value to the denominator. For a gene with variances stored in the array \code{x}, add \code{quantile(x, quantileOffset)} before taking the reciprocal.}

\item{maxRatio}{When computing precision as the reciprocal of variance \code{1/(x+tau)} select tau to have a maximum ratio between the largest and smallest precision}

\item{h5adBlockSizes}{set the automatic block size block size (in bytes) for DelayedArray to read an H5AD file.  Larger values use more memory but are faster.}

\item{verbose}{Show messages, defaults to TRUE}
}
\description{
Compute precision weights for pseudobulk using the delta method to approximate the variance of the log2 counts per million considering variation in the number of cells and gene expression variance across cells within each sample. By default, used number of cells; if specified use delta method.  Note that \code{processAssays()} uses number of cells as weights when no weights are specificed
}
\examples{
library(muscat)

data(example_sce)

# create pseudobulk for each sample and cell cluster
pb <- aggregateToPseudoBulk(example_sce,
  assay = "counts",
  sample_id = "sample_id",
  cluster_id = "cluster_id",
  verbose = FALSE
)

# Create precision weights for pseudobulk
# By default, weights are set to cell count, 
# which is the default in processAssays() 
# even when no weights are specified
weightsList <- pbWeights(example_sce,
    sample_id = "sample_id",
    cluster_id = "cluster_id")

# voom-style normalization using initial weights
res.proc <- processAssays(pb, ~ group_id, weightsList = weightsList)
}
