% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotForest.R
\name{plotForest}
\alias{plotForest}
\title{Forest plot}
\usage{
plotForest(x, coef, gene)
}
\arguments{
\item{x}{result from \code{dreamlet}}

\item{coef}{coefficient to test with \code{topTable}}

\item{gene}{gene to show results for}
}
\description{
Forest plot
}
