% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVarPart.R
\name{plotVarPart,DataFrame-method}
\alias{plotVarPart,DataFrame-method}
\alias{plotVarPart,dreamletProcessedData,dreamletProcessedData-method}
\title{Violin plot of variance fractions}
\usage{
\S4method{plotVarPart}{DataFrame}(
  obj,
  col = c(ggColorHue(base::ncol(obj) - 3), "grey85"),
  label.angle = 20,
  main = "",
  ylab = "",
  convertToPercent = TRUE,
  ...
)
}
\arguments{
\item{obj}{\code{varParFrac} object returned by \code{fitExtractVarPart} or \code{extractVarPart}}

\item{col}{vector of colors}

\item{label.angle}{angle of labels on x-axis}

\item{main}{title of plot}

\item{ylab}{text on y-axis}

\item{convertToPercent}{multiply fractions by 100 to convert to percent values}

\item{...}{additional arguments}
}
\description{
Violin plot of variance fraction for each gene and each variable
}
