% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotZenithResults.R
\name{plotZenithResults}
\alias{plotZenithResults}
\title{Heatmap of zenith results}
\usage{
plotZenithResults(df, ntop = 5, nbottom = 5)
}
\arguments{
\item{df}{result \code{data.frame} from \link{zenith_gsa}}

\item{ntop}{number of gene sets with highest t-statistic to show}

\item{nbottom}{number of gene sets with lowest t-statistic to show}
}
\value{
Heatmap showing enrichment for gene sets and cell types
}
\description{
Heatmap of zenith results showing genesets that have the top and bottom t-statistics from each assay.
}
\examples{
library(muscat)
library(SingleCellExperiment)

data(example_sce)

# create pseudobulk for each sample and cell cluster
pb <- aggregateToPseudoBulk(example_sce, 
   assay = "counts",    
   cluster_id = 'cluster_id', 
   sample_id = 'sample_id',
   verbose=FALSE)

# voom-style normalization
res.proc = processAssays( pb, ~ group_id)

# Differential expression analysis within each assay,
# evaluated on the voom normalized data 
res.dl = dreamlet( res.proc, ~ group_id)

# Load Gene Ontology database 
# use gene 'SYMBOL', or 'ENSEMBL' id
# use get_MSigDB() to load MSigDB
library(zenith)
go.gs = get_GeneOntology("CC", to="SYMBOL")
   
# Run zenith gene set analysis on result of dreamlet
res_zenith = zenith_gsa(res.dl, coef = 'group_idstim', go.gs)

# for each cell type select 5 genesets with largest t-statistic
# and 1 geneset with the lowest
# Grey boxes indicate the gene set could not be evaluted because
#    to few genes were represented
plotZenithResults(res_zenith, 5, 1)

}
