% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamlet.R
\name{print,dreamletProcessedData-method}
\alias{print,dreamletProcessedData-method}
\alias{print,dreamletProcessedData,dreamletProcessedData-method}
\title{Print object}
\usage{
\S4method{print}{dreamletProcessedData}(x, ...)
}
\arguments{
\item{x}{dreamletProcessedData object}

\item{...}{other arguments}
}
\description{
Print object
}
