% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processAssays.R
\name{processAssays}
\alias{processAssays}
\title{Processing SingleCellExperiment to dreamletProcessedData}
\usage{
processAssays(
  sceObj,
  formula,
  assays = assayNames(sceObj),
  min.cells = 10,
  min.count = 10,
  min.samples = 4,
  min.prop = 0.4,
  isCounts = TRUE,
  normalize.method = "TMM",
  useCountsWeights = TRUE,
  pmetadata = NULL,
  pkeys = NULL,
  quiet = FALSE,
  BPPARAM = SerialParam(),
  ...
)
}
\arguments{
\item{sceObj}{SingleCellExperiment object}

\item{formula}{regression formula for differential expression analysis}

\item{assays}{array of assay names to include in analysis. Defaults to \code{assayNames(sceObj)}}

\item{min.cells}{minimum number of observed cells for a sample to be included in the analysis}

\item{min.count}{minimum number of reads for a gene to be considered expressed in a sample.  Passed to \code{edgeR::filterByExpr}}

\item{min.samples}{minimum number of samples passing cutoffs for cell cluster to be retained}

\item{min.prop}{minimum proportion of retained samples with non-zero counts for a gene to be retained}

\item{isCounts}{logical, indicating if data is raw counts}

\item{normalize.method}{normalization method to be used by \code{calcNormFactors}}

\item{useCountsWeights}{use cell count weights}

\item{pmetadata}{sample-specific metadata the varies across cell types.  This is merged with \code{colData(sceObj)} for each assay to make variables accessable to the formula}

\item{pkeys}{array of two strings indicating sample identifier and cell type identifier columns in pmetadata}

\item{quiet}{show messages}

\item{BPPARAM}{parameters for parallel evaluation}

\item{...}{other arguments passed to \code{dream}}
}
\value{
Object of class \code{dreamletProcessedData} storing voom-style normalized expression data
}
\description{
For raw counts, estimate precision weights using linear mixed model weighting by number of cells observed for each sample.  For normalized data, only weight by number of cells.
}
\details{
For each cell cluster, samples with at least \code{min.cells} are retained. Only clusters with at least \code{min.samples} retained samples are kept. Genes are retained if they have at least \code{min.count} reads in at least \code{min.prop} fraction of the samples.  Current values are reasonable defaults, since genes that don't pass these cutoffs are very underpowered for differential expression analysis and only increase the multiple testing burden.  But values of \code{min.cells = 5} and \code{min.count = 5} are also reasonable if you want to include more genes in the analysis.

The precision weights are estimated using the residuals fit from the specified formula.  These weights are robust to changes in the formula as long as the major variables explaining the highest fraction of the variance are included.
}
\examples{
library(muscat)
library(SingleCellExperiment)

data(example_sce)

# create pseudobulk for each sample and cell cluster
pb <- aggregateToPseudoBulk(example_sce, 
   assay = "counts",    
   cluster_id = 'cluster_id', 
   sample_id = 'sample_id',
   verbose=FALSE)

# voom-style normalization
res.proc = processAssays( pb, ~ group_id)

# Differential expression analysis within each assay,
# evaluated on the voom normalized data 
res.dl = dreamlet( res.proc, ~ group_id)

}
