% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamlet.R
\name{processAssays}
\alias{processAssays}
\title{Processing SingleCellExperiment to dreamletProcessedData}
\usage{
processAssays(
  sceObj,
  formula,
  min.cells = 10,
  isCounts = TRUE,
  normalize.method = "TMM",
  BPPARAM = bpparam(),
  ...
)
}
\arguments{
\item{sceObj}{SingleCellExperiment object}

\item{formula}{regression formula for differential expression analysis}

\item{min.cells}{minimum number of observed cells for a sample to be included in the analysis}

\item{isCounts}{logical, indicating if data is raw counts}

\item{normalize.method}{normalization method to be used by \code{calcNormFactors}}

\item{BPPARAM}{parameters for parallel evaluation}

\item{...}{other arguments passed to \code{dream}}

\item{data}{metadata used in regression formula}
}
\description{
For raw counts, estimate precision weights using linear mixed model weighting by number of cells observed for each sample.  For normalized data, only weight by number of cells
}
