% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamlet.R
\name{processOneAssay}
\alias{processOneAssay}
\title{Processing expression data from assay}
\usage{
processOneAssay(
  y,
  formula,
  data,
  min.cells = 10,
  normalize.method = "TMM",
  BPPARAM = bpparam(),
  ...
)
}
\arguments{
\item{y}{matrix of counts or log2 CPM}

\item{formula}{regression formula for differential expression analysis}

\item{data}{metadata used in regression formula}

\item{min.cells}{minimum number of observed cells for a sample to be included in the analysis}

\item{normalize.method}{normalization method to be used by \code{calcNormFactors}}

\item{BPPARAM}{parameters for parallel evaluation}

\item{...}{other arguments passed to \code{dream}}
}
\description{
For raw counts, estimate precision weights using linear mixed model weighting by number of cells observed for each sample.  For normalized data, only weight by number of cells
}
