% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeConstantTerms.R
\name{removeConstantTerms}
\alias{removeConstantTerms}
\title{Remove constant terms from formula}
\usage{
removeConstantTerms(formula, data)
}
\arguments{
\item{formula}{original formula}

\item{data}{data.frame}
}
\description{
Remove constant terms from formula
}
