% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamlet.R, R/dreamletProcessedData.R
\name{show,dreamletResult-method}
\alias{show,dreamletResult-method}
\alias{show,dreamletResult,dreamletResult-method}
\alias{show,dreamletProcessedData-method}
\alias{show,dreamletProcessedData,dreamletProcessedData-method}
\title{Show object}
\usage{
\S4method{show}{dreamletResult}(object)

\S4method{show}{dreamletProcessedData}(object)
}
\arguments{
\item{object}{\code{dreamletProcessedData} object}
}
\value{
show data stored in object
}
\description{
Show object

Show object
}
