% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash.R
\name{tabToMatrix}
\alias{tabToMatrix}
\title{Show object
Convert results table to matrix}
\usage{
tabToMatrix(tab, col, rn, cn)
}
\arguments{
\item{tab}{results table from \code{topTable()}}

\item{col}{which column to extract}

\item{rn}{rownames}

\item{cn}{colnames}
}
\description{
Convert results table to matrix
}
