% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenith_gsa.R
\name{zenith_gsa}
\alias{zenith_gsa}
\alias{zenith_gsa,dreamletResult,ANY,GeneSetCollection-method}
\title{Perform zenith analysis}
\usage{
zenith_gsa(
  fit,
  coefs,
  geneSets,
  n_genes_min = 10,
  inter.gene.cor = NA,
  progressbar = TRUE
)

\S4method{zenith_gsa}{dreamletResult,ANY,GeneSetCollection}(
  fit,
  coefs,
  geneSets,
  n_genes_min = 10,
  inter.gene.cor = NA,
  progressbar = TRUE
)
}
\arguments{
\item{fit}{results from \code{dreamlet}}

\item{coefs}{coefficients to test using \code{topTable(fit, coef)}}

\item{geneSets}{a list of gene sets}

\item{n_genes_min}{minumum number of genes in a geneset}

\item{inter.gene.cor}{if NA, estimate correlation from data.  Otherwise, use specified value}

\item{progressbar}{if TRUE, show progress bar}
}
\description{
Perform zenith analysis
}
