% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenith_gsa.R
\name{zenith_gsa}
\alias{zenith_gsa}
\title{zenith gene set analysis on dreamlet results}
\usage{
zenith_gsa(res.dl, coef, geneSets, n_genes_min = 10)
}
\arguments{
\item{res.dl}{results from \code{dreamlet}}

\item{coef}{coefficient to test using \code{topTable(fit, coef)}}

\item{geneSets}{a list of gene sets}

\item{n_genes_min}{minumum number of genes in a geneset}
}
\description{
zenith gene set analysis on dreamlet results
}
